[WIP] Add I2C driver & I2C manager

+++ drivers/lcd_rk043fn48h.hpp
@@ -21,9 +21,9 @@ namespace drivers
 class glcd_rk043fn48h : public hal::interface::glcd<uint16_t>
 {
 private:
-    static constexpr size_t width_px = 480;
-    static constexpr size_t height_px = 272;
-    static constexpr size_t bits_per_px = 5 + 6 + 5; // RGB565
+    static constexpr uint16_t width_px = 480;
+    static constexpr uint16_t height_px = 272;
+    static constexpr uint8_t bits_per_px = 5 + 6 + 5; // RGB565
 
 public:
     using pixel_t = pixel_t;
@@ -33,9 +33,9 @@ public:
     glcd_rk043fn48h(const std::array<const gpio::io, 29> &gpios, framebuffer_t &frame_buffer, bool portrait_mode = false);
     ~glcd_rk043fn48h();
 
-    size_t width(void) override { return this->portrait_mode ? height_px : width_px; }
-    size_t height(void) override { return this->portrait_mode ? width_px : height_px; }
-    size_t bpp(void) override { return bits_per_px; }
+    uint16_t width(void) override { return this->portrait_mode ? height_px : width_px; }
+    uint16_t height(void) override { return this->portrait_mode ? width_px : height_px; }
+    uint8_t bpp(void) override { return bits_per_px; }
 
     void draw_pixel(int16_t x, int16_t y, pixel_t pixel) override;
     void draw_data(int16_t x0, int16_t y0, int16_t x1, int16_t y1, pixel_t *data) override;
